Update for generation step
